+++ main.py
@@ -1,6 +1,15 @@
+from time import sleep
+
 import app
-from common.util import print_header
+from common import eop
+from common.util import print_header, vert_center, module_loaded
 
 if __name__ == '__main__':
-    print_header("Welcome to Muninn")
+
+    if not eop.is_root():
+        vert_center("Please Run as Root. Required for scanner")
+    else:
+        module_loaded("Welcome to Muninn", sep="=")
+
+
     app.run()