+++ modules/Scanning/Scan.py
@@ -7,9 +7,10 @@ def Controller(scanner=None):
 
     if scanner is None:
         module_loaded("New Scan")
+        scanner = Scanner(hosts="127.0.0.1", ports=[22], args="-sV -O")
+        print("Default Command: " + scanner.current_command)
     else:
         module_loaded("Scan with " + scanner.current_command)
-    scanner = Scanner(hosts="127.0.0.1", ports=[22])
     actions = {"Configure": scanner.Configure, "Run": scanner.Run}
     selection = options(actions, "Select", "Select an action")
     if not selection: